extract StyledMdPreview to prevent circular import
